Move loginMehtodCreater (firebase-auth-own-ui.tsx -> input-field-utils)
